fix: 2d 지도 초기화 함수 분리
